backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-docx
jinja2
requests
aiofiles
sqlalchemy
alembic
pydantic
python-dotenv
